Plots can be saved to the function output
The two arguments, "save.dist.plots" and "save.err.plots" have been added, which allow for saving distribution and error plots to the function output, in addition to plotting them in the graphics device. An efficient alternative is to set  "save.data=TRUE" and leave all plotting as "FALSE". The resulting data frame in the output list contains all the values that are used in the function plots, therefore, it is possible to make your own plots.

All plotting is now set to FALSE by default. It can be slow for large datasets. Another option is to us the "sample" argument to subsample large backscatter rasters.

The statistics provided in the output now describe the error prior to the bulk shift, in addition to the residual error, making it possible to compare the amount of error reduction provided by the model.